attr: Fix the ptest path to test-driver
Add a sed substitution for ${PTEST_PATH}/Makefile that transforms
 TEST_LOG_DRIVER = $(SHELL) $(top_srcdir)/build-aux/test-driver
to
 TEST_LOG_DRIVER = $(SHELL) $(top_builddir)/build-aux/test-driver
which is where the test-driver script can be found when installed.

There used to be an oe-core automake patch to do this, but it
broke non-ptest use of automake.

Relates to [YOCTO #15635]

(From OE-Core rev: 289db96b441e01c4ba1bbcd76fcf52b3258f2613)

Signed-off-by: Ola x Nilsson <olani@axis.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>