libtool: update 2.5.2 -> 2.5.4
2.5.2 was a beta, upstream made two stable releases since:
https://savannah.gnu.org/news/?id=10676
https://savannah.gnu.org/news/?id=10693

Drop libtool/0001-ltmain.in-Handle-trailing-slashes-on-install-command.patch
(merged upstream)

Apply the multilib patch unconditionally as it could be applied to add variants.

For the other it was pointed out that it:
"potentially has side effects for native builds. It breaks the ability
to run files in the .libs directory which in the cross case we don't
care about but some natives did once used to do that. The risk is they
find host things instead."

License-Update: change of FSF address to a URI

(From OE-Core rev: 9a672faf21648c60d7ff29439bb236f3ac87100d)

Signed-off-by: Alexander Kanavin <alex@linutronix.de>
Signed-off-by: Mathieu Dubois-Briand <mathieu.dubois-briand@bootlin.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>